Fully separated abstract class Board, comments, JavaDoc, minor changes.
- PlayerBoard and ComputerBoard are separated from Board abstract class to a high extent
- Revised Ship abstract class and all extentions from it
- More accessor and mutator methods in Board class
- JavaDoc and comments
- Numbers and characters added to the board
