--- battleship/byKN/AircraftCarrier.java
@@ -1,11 +1,0 @@
-
-public class AircraftCarrier extends Ship{
-	
-	
-	public AircraftCarrier()
-	{
-		super(5, "Aircraft Carrier");
-		
-	}
-
-}
