--- battleship/byKN/Battleship.java
@@ -1,10 +1,0 @@
-
-public class Battleship extends Ship{
-
-	
-	public Battleship()
-	{
-		super(4, "Battleship");
-		
-	}
-}
