--- battleship/byKN/Cruiser.java
@@ -1,12 +1,0 @@
-
-public class Cruiser extends Ship{
-	
-	
-	
-	public Cruiser()
-	{
-		super(3, "Cruiser");	
-		
-	}
-
-}
