--- battleship/byKN/Destroyer.java
@@ -1,10 +1,0 @@
-
-public class Destroyer extends Ship{
-
-		
-	public Destroyer()
-	{
-		super(2, "Destroyer");	
-		
-	}
-}
