--- battleship/byKN/Submarine.java
@@ -1,11 +1,0 @@
-
-public class Submarine extends Ship{
-
-	
-	
-	public Submarine()
-	{
-		super(3, "Submarine");	
-		
-	}
-}
